Split parsing out into separate file
